refactor: Rename CustumMovmentScene to QarakusiScene, move it to QarakusiScene and changes in related files
