+++ lib/movement_problems/__init__.py
@@ -1,5 +1,4 @@
 from lib.movement_problems.coordinate import Coordinate
 from lib.movement_problems.helper_functions import obj_movement, down_brace_with_text, up_brace_with_text
-from lib.movement_problems.custom_movement_scene import CustomMovementScene
 
-__all__ = ['down_brace_with_text', 'obj_movement', 'Coordinate', 'CustomMovementScene', 'up_brace_with_text']
+__all__ = ['down_brace_with_text', 'obj_movement', 'Coordinate', 'up_brace_with_text']
